Closes #419  
Closes #401  
~~It is just the main idea. But before continuing I would like some comment.~~

## [Hello World benchmark](https://github.com/botika/buf-min/blob/e62a7975ed7ea00c473902ae774af537c0374949/benches/src/all.rs#L122-L131)
### Old
```
Buffer Bytes            time:   [30.068 ns 30.155 ns 30.273 ns] 
```
### New
```
Buffer Bytes            time:   [14.207 ns 14.366 ns 14.573 ns]                          
                        change: [-51.621% -50.128% -48.373%] (p = 0.00 < 0.05)
                        Performance has improved.
```