@seanmonstar 2020-07-20
What's the motivation for handling allocation ourselves instead of reusing what `Vec` does?