@botika 2020-07-20
Win on these benchmarks https://github.com/Kogia-sima/sailfish/tree/master/benches#results