@botika 2020-07-21
With these changes it has been achieved:
- Avoid double checks when we use `BytesMut.reserve(usize)`
- Improve `rustc` inline optimizations
- "Real" zero cost `BytesMut.freeze()`, resolve #418