@botika 2020-07-21
@seanmonstar In the benchmarks, https://github.com/botika/buf-min/tree/e62a7975ed7ea00c473902ae774af537c0374949/benches, you can see the difference. Just change the byte version in `Cargo.toml`. In conclusion, for a `"Hello World!"`, it takes half the time. 
I can add all the coverage that you think is convenient to have more soundness.