@seanmonstar 2020-07-21
The difference is very cool! Would you be able to explain the differences proposed here, vs what `Vec` is doing? Usually, just relying on `std` is a good idea, since they handle safety and speed. But if we are going to differ, we should be able to explain exactly why.