@botika 2020-07-21
> The difference is very cool! Would you be able to explain the differences proposed here, vs what `Vec` is doing? Usually, just relying on `std` is a good idea, since they handle safety and speed. But if we are going to differ, we should be able to explain exactly why.

- Avoid this check https://github.com/rust-lang/rust/blob/8ad7bc3f428300aee6764f6e23527e19eb235e81/src/liballoc/raw_vec.rs#L309
Since previously we have already done it.
- Minimize this issue, https://github.com/rust-lang/rust/issues/74020 
Every structure will always be initialized. In our case, `Vec` and` RawVec` for nothing and several times.
- Optimize inline-able generated code
`BytesMut::with_capacity` always generates less instructions
- Zero cost freeze 
Not depending on `Vec` I can be sure that freeze does not relocate
- Very simple implementation
Does not require maintenance and is specific to the problem
- Fastest