@fanatid 2020-07-21
@botika can you comment this? Should be vector be shrinked?
https://github.com/tokio-rs/bytes/pull/418#issuecomment-660846397