@botika 2020-07-21
@fanatid I don't see any capacity in the free call, I do not know what he's talking about.
https://www.tutorialspoint.com/c_standard_library/c_function_free.htm