@Darksonn 2020-07-21
When running the following program with miri, it fails due to the memory being deallocated with the wrong capacity.
```toml
[package]
name = "dealloc-test"
version = "0.1.0"
authors = ["Alice Ryhl <alice@ryhl.io>"]
edition = "2018"

[dependencies]
bytes = { git = "https://github.com/botika/bytes", rev = "alloc" }
```
```rust
use bytes::BytesMut;

fn main() {
    let mut b = BytesMut::with_capacity(16);
    b.extend_from_slice(&[0; 8]);
    let frozen = b.freeze();
    drop(frozen);
}
```
```
error: Undefined Behavior: incorrect layout on deallocation: allocation has size 16 and alignment 1, but gave size 8 and alignment 1
   --> /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/liballoc/alloc.rs:102:14
    |
102 |     unsafe { __rust_dealloc(ptr, layout.size(), layout.align()) }
    |              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ incorrect layout on deallocation: allocation has size 16 and alignment 1, but gave size 8 and alignment 1
    |
    = help: this indicates a bug in the program: it performed an invalid operation, and caused Undefined Behavior
    = help: see https://doc.rust-lang.org/nightly/reference/behavior-considered-undefined.html for further information
            
    = note: inside `std::alloc::dealloc` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/liballoc/alloc.rs:102:14
    = note: inside `<std::alloc::Global as std::alloc::AllocRef>::dealloc` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/liballoc/alloc.rs:186:22
    = note: inside `alloc::alloc::box_free::<[u8]>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/liballoc/alloc.rs:295:9
    = note: inside `std::intrinsics::drop_in_place::<std::boxed::Box<[u8]>> - shim(Some(std::boxed::Box<[u8]>))` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libcore/ptr/mod.rs:184:1
    = note: inside `std::mem::drop::<std::boxed::Box<[u8]>>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libcore/mem/mod.rs:883:24
    = note: inside closure at /home/alice/.cargo/git/checkouts/bytes-31cde87f5a385f88/03f963b/src/bytes.rs:903:13
    = note: inside `<std::sync::atomic::AtomicPtr<()> as bytes::loom::sync::atomic::AtomicMut<()>>::with_mut::<[closure@bytes::bytes::promotable_even_drop::{{closure}}#0 0:&*const u8, 1:&usize], ()>` at /home/alice/.cargo/git/checkouts/bytes-31cde87f5a385f88/03f963b/src/loom.rs:17:17
    = note: inside `bytes::bytes::promotable_even_drop` at /home/alice/.cargo/git/checkouts/bytes-31cde87f5a385f88/03f963b/src/bytes.rs:894:5
    = note: inside `<bytes::Bytes as std::ops::Drop>::drop` at /home/alice/.cargo/git/checkouts/bytes-31cde87f5a385f88/03f963b/src/bytes.rs:525:18
    = note: inside `std::intrinsics::drop_in_place::<bytes::Bytes> - shim(Some(bytes::Bytes))` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libcore/ptr/mod.rs:184:1
    = note: inside `std::mem::drop::<bytes::Bytes>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libcore/mem/mod.rs:883:24
note: inside `main` at src/main.rs:7:5
   --> src/main.rs:7:5
    |
7   |     drop(frozen);
    |     ^^^^^^^^^^^^
    = note: inside closure at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/rt.rs:67:34
    = note: inside closure at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/rt.rs:52:73
    = note: inside `std::sys_common::backtrace::__rust_begin_short_backtrace::<[closure@std::rt::lang_start_internal::{{closure}}#0::{{closure}}#0 0:&dyn std::ops::Fn() -> i32 + std::marker::Sync + std::panic::RefUnwindSafe], i32>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/sys_common/backtrace.rs:130:5
    = note: inside closure at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/rt.rs:52:13
    = note: inside `std::panicking::r#try::do_call::<[closure@std::rt::lang_start_internal::{{closure}}#0 0:&&dyn std::ops::Fn() -> i32 + std::marker::Sync + std::panic::RefUnwindSafe], i32>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/panicking.rs:348:40
    = note: inside `std::panicking::r#try::<i32, [closure@std::rt::lang_start_internal::{{closure}}#0 0:&&dyn std::ops::Fn() -> i32 + std::marker::Sync + std::panic::RefUnwindSafe]>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/panicking.rs:325:15
    = note: inside `std::panic::catch_unwind::<[closure@std::rt::lang_start_internal::{{closure}}#0 0:&&dyn std::ops::Fn() -> i32 + std::marker::Sync + std::panic::RefUnwindSafe], i32>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/panic.rs:394:14
    = note: inside `std::rt::lang_start_internal` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/rt.rs:51:25
    = note: inside `std::rt::lang_start::<()>` at /home/alice/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/rt.rs:67:5

error: aborting due to previous error
```